requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-docx
sentence-transformers
pytesseract
Pillow